requirements.txt
# He'story Backend Requirements
# 이 파일은 Railway 배포를 위한 것입니다. 실제 requirements는 backend/requirements.txt를 참조하세요.

fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
google-generativeai==0.3.2
google-genai==0.1.0
# pyaudio==0.2.14  # Railway 배포에서는 오디오 기능 비활성화
aiofiles==23.2.1
python-dotenv==1.0.0
httpx==0.26.0
websockets==12.0